Move all cisco specific message handling off of telnet socket Add correct poe enabled flag support handling Add better failure/skip test messaging Miscellaneous code cleanup
